requirements.txt
typing-extensions==3.7.4.3
uvloop==0.15.0
aiohttp[speedups]==3.8.1
pydantic==1.8.2
loguru==0.5.3
python-dotenv==0.19.0
ujson==4.0.2
tortoise-orm[asyncpg]==0.17.8
typer[all]==0.4.0
psycopg2-binary==2.9.2
aerich==0.5.8
aiocron==1.8
aiogram==2.16
aiogram-dialog==1.1.0
